# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Acadex API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ChallengeDifficulty =
  (typeof ChallengeDifficulty)[keyof typeof ChallengeDifficulty];

export const ChallengeDifficulty = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
} as const;

export interface Challenge {
  id: number;
  title: string;
  description: string;
  subject: string;
  difficulty: ChallengeDifficulty;
  authorName: string;
  createdAt: string;
  responseCount: number;
}

export interface Response {
  id: number;
  challengeId: number;
  content: string;
  authorName: string;
  createdAt: string;
}

export type ChallengeDetail = Challenge & {
  responses: Response[];
};

export type CreateChallengeBodyDifficulty =
  (typeof CreateChallengeBodyDifficulty)[keyof typeof CreateChallengeBodyDifficulty];

export const CreateChallengeBodyDifficulty = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
} as const;

export interface CreateChallengeBody {
  title: string;
  description: string;
  subject: string;
  difficulty: CreateChallengeBodyDifficulty;
  authorName: string;
}

export interface CreateResponseBody {
  content: string;
  authorName: string;
}

export type ChallengeStatsBySubjectItem = {
  subject: string;
  count: number;
};

export type ChallengeStatsByDifficultyItem = {
  difficulty: string;
  count: number;
};

export interface ChallengeStats {
  totalChallenges: number;
  totalResponses: number;
  bySubject: ChallengeStatsBySubjectItem[];
  byDifficulty: ChallengeStatsByDifficultyItem[];
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  challenge: "challenge",
  response: "response",
} as const;

export interface ActivityItem {
  type: ActivityItemType;
  id: number;
  challengeId?: number;
  title: string;
  subject: string;
  authorName: string;
  createdAt: string;
}

export interface ErrorResponse {
  error: string;
}

export type ListChallengesParams = {
  subject?: string;
  difficulty?: string;
};