# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Acadex API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: challenges
    description: Academic challenge operations
  - name: responses
    description: Response operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /challenges:
    get:
      operationId: listChallenges
      tags: [challenges]
      summary: List all challenges
      parameters:
        - name: subject
          in: query
          schema:
            type: string
        - name: difficulty
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of challenges
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Challenge"
    post:
      operationId: createChallenge
      tags: [challenges]
      summary: Post a new challenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChallengeBody"
      responses:
        "201":
          description: Created challenge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Challenge"

  /challenges/{id}:
    get:
      operationId: getChallenge
      tags: [challenges]
      summary: Get a challenge by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Challenge detail with responses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengeDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /challenges/{id}/responses:
    post:
      operationId: createResponse
      tags: [responses]
      summary: Post a response to a challenge
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResponseBody"
      responses:
        "201":
          description: Created response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"

  /challenges/stats/summary:
    get:
      operationId: getChallengeStats
      tags: [challenges]
      summary: Get summary stats (counts by subject and difficulty)
      responses:
        "200":
          description: Challenge stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengeStats"

  /challenges/recent/activity:
    get:
      operationId: getRecentActivity
      tags: [challenges]
      summary: Get recent challenge and response activity
      responses:
        "200":
          description: Recent activity feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    Challenge:
      type: object
      required: [id, title, description, subject, difficulty, authorName, createdAt, responseCount]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        subject:
          type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        authorName:
          type: string
        createdAt:
          type: string
          format: date-time
        responseCount:
          type: integer

    ChallengeDetail:
      allOf:
        - $ref: "#/components/schemas/Challenge"
        - type: object
          required: [responses]
          properties:
            responses:
              type: array
              items:
                $ref: "#/components/schemas/Response"

    CreateChallengeBody:
      type: object
      required: [title, description, subject, difficulty, authorName]
      properties:
        title:
          type: string
        description:
          type: string
        subject:
          type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        authorName:
          type: string

    Response:
      type: object
      required: [id, challengeId, content, authorName, createdAt]
      properties:
        id:
          type: integer
        challengeId:
          type: integer
        content:
          type: string
        authorName:
          type: string
        createdAt:
          type: string
          format: date-time

    CreateResponseBody:
      type: object
      required: [content, authorName]
      properties:
        content:
          type: string
        authorName:
          type: string

    ChallengeStats:
      type: object
      required: [totalChallenges, totalResponses, bySubject, byDifficulty]
      properties:
        totalChallenges:
          type: integer
        totalResponses:
          type: integer
        bySubject:
          type: array
          items:
            type: object
            required: [subject, count]
            properties:
              subject:
                type: string
              count:
                type: integer
        byDifficulty:
          type: array
          items:
            type: object
            required: [difficulty, count]
            properties:
              difficulty:
                type: string
              count:
                type: integer

    ActivityItem:
      type: object
      required: [type, id, title, subject, authorName, createdAt]
      properties:
        type:
          type: string
          enum: [challenge, response]
        id:
          type: integer
        challengeId:
          type: integer
        title:
          type: string
        subject:
          type: string
        authorName:
          type: string
        createdAt:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string