# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Acadex API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all challenges
 */
export const ListChallengesQueryParams = zod.object({
  subject: zod.coerce.string().optional(),
  difficulty: zod.coerce.string().optional(),
});

export const ListChallengesResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  subject: zod.string(),
  difficulty: zod.enum(["beginner", "intermediate", "advanced"]),
  authorName: zod.string(),
  createdAt: zod.coerce.date(),
  responseCount: zod.number(),
});
export const ListChallengesResponse = zod.array(ListChallengesResponseItem);

/**
 * @summary Post a new challenge
 */
export const CreateChallengeBody = zod.object({
  title: zod.string(),
  description: zod.string(),
  subject: zod.string(),
  difficulty: zod.enum(["beginner", "intermediate", "advanced"]),
  authorName: zod.string(),
});

/**
 * @summary Get a challenge by ID
 */
export const GetChallengeParams = zod.object({
  id: zod.coerce.number(),
});

export const GetChallengeResponse = zod
  .object({
    id: zod.number(),
    title: zod.string(),
    description: zod.string(),
    subject: zod.string(),
    difficulty: zod.enum(["beginner", "intermediate", "advanced"]),
    authorName: zod.string(),
    createdAt: zod.coerce.date(),
    responseCount: zod.number(),
  })
  .and(
    zod.object({
      responses: zod.array(
        zod.object({
          id: zod.number(),
          challengeId: zod.number(),
          content: zod.string(),
          authorName: zod.string(),
          createdAt: zod.coerce.date(),
        }),
      ),
    }),
  );

/**
 * @summary Post a response to a challenge
 */
export const CreateResponseParams = zod.object({
  id: zod.coerce.number(),
});

export const CreateResponseBody = zod.object({
  content: zod.string(),
  authorName: zod.string(),
});

/**
 * @summary Get summary stats (counts by subject and difficulty)
 */
export const GetChallengeStatsResponse = zod.object({
  totalChallenges: zod.number(),
  totalResponses: zod.number(),
  bySubject: zod.array(
    zod.object({
      subject: zod.string(),
      count: zod.number(),
    }),
  ),
  byDifficulty: zod.array(
    zod.object({
      difficulty: zod.string(),
      count: zod.number(),
    }),
  ),
});

/**
 * @summary Get recent challenge and response activity
 */
export const GetRecentActivityResponseItem = zod.object({
  type: zod.enum(["challenge", "response"]),
  id: zod.number(),
  challengeId: zod.number().optional(),
  title: zod.string(),
  subject: zod.string(),
  authorName: zod.string(),
  createdAt: zod.coerce.date(),
});
export const GetRecentActivityResponse = zod.array(
  GetRecentActivityResponseItem,
);